Fix bug on the addPath method
Create the path folder as a nested folder of the parentURL folder
instead of creating it as a nested folder of the data folder.

Parse the parentURL and the path to make sure that the path starts with
a backslash and that the parentURL doesn't end with a backslash,
otherwise, the metadata file will have incorrect URLs.

On the addPath method unit test, test for paths that starts with or
without a backslash and for URLs that ends with or without a backslash.