Make the crawler works
Add the function crawl.js to crawl for new URLs.

On index.js add a recursive function to keep calling the crawl function
if happens an error.

Add axios to the project to make the get request to the URLs and create
the npm script "start" as an alias for "node index.js".